Add `src/` directory to webpack's resolve.root
Adding the `src/` directory to resolve.root allows developers to import
modules without having to give the relative path to the current file
imoprting the module.  This makes it easier to refactor the code as it's
easier to reason about module paths relative to the project's root
folder.

This is also backwards compatible, and relative imports are still
possible.

Closes #66